+++ python-bindings/b2_terraform/provider_tool.py
@@ -447,6 +447,7 @@ class AccountInfo(Command):
             'accountAuthToken': account_info.get_account_auth_token(),
             'apiUrl': account_info.get_api_url(),
             'downloadUrl': account_info.get_download_url(),
+            's3_ApiUrl': account_info.get_s3_api_url(),
         }